==> Instacart Grocery Analysis/03 Scripts/1. Task 4.2 InstaCart Libraries and Python Data Types.ipynb <==
# 1. IMPORT LIBRARIES

import pandas as pd #importing pandas library
import numpy as np #importing NumPy library
import os #importing os

# 2. PYTHON DATA TYPES


## Code 3 different ways of reaching a result of 100 by adding or subtracting numeric variables.

### Substraction

x = 150
y = 50

x - y

### Addition

x = 50
y = 25
z = 25

x + y + z

### Combination of addition and substraction

x = 200
y = 50
z = 50

x-(y + z)

## Code 2 floating-point variables and divide them by each other.

# Define two floating-point variables
a = 40.234
b = 2.5

# Divide the first variable by the second
a/b

## Construct a short word made of separate strings 

c = 'Python '
d = 'and Jupyter'

c + d

## Construct 2 short sentences made of separate strings.

#Define two variables to store string
e = 'Anaconda is a file- and library-management system '
f = ' and Jupyter is the online notebook used to write and run the Python code.'

e + f

# 3. BONUS TASK

# Write a command that makes a string appear 100 times.
e  + f * 100
