==> Instacart Grocery Analysis/03 Scripts/2.  Task 4.3 InstaCart Data import and descriptive analysis.ipynb <==
# 01. Import Libraries

import pandas as pd
import numpy as np
import os

# 02. Import the Data

df = pd.read_csv(r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4/02 Data/Original Data/orders.csv',index_col = False)

## 1. Turn your project folder path into a string

path = r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4'

## Display the result of folder path into variable 'path'

path

## 2. Store the list of columns into a variable, , omitting the “eval_set” column in your final dataframe.

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

## 3. Import the “orders.csv” data set into your notebook using the os library and the vars_list shortcut.

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list, index_col= False) 

## 4. Print the first 5 rows of your dataframe.

df.head()

## 5. Import the “products.csv” file into Jupyter as df_prods.

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

## 6. Print the first 20 rows of the df_prods dataframe.

df_prods.head(25)

## 7. print the last 35 rows of df_prods dataframe. 

df_prods.tail(35)

## 8. Print the names of the columns in the df_prods dataframe.

df_prods.columns

## 9. Print the number of rows and columns in the df_prods dataframe.

df_prods.shape

# 03. Answer the following questions using the relevant commands.

## 1. What’s the maximum value of the “aisle_id” column?

df_prods.describe()

### Ans: Max value of 'aisle_id' is 134.

## 2. What’s the data type of the “department_id” column?

df_prods.dtypes

### Ans. Data type of "department_id" is Integer.
