==> Instacart Grocery Analysis/03 Scripts/3. Exercise & Task 4.4.ipynb <==
# EXERCISE
# 01. Import Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

# 02. Create the folder path

# Turn the folder path into variable to get access easily. 
path = r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4'

# 03. Import the data into variables

# Import 'order.csv' data set from original Data folder.
df_ords = pd.read_csv(os.path.join(path, '02 Data','Original Data','orders.csv'), index_col = False)

# Import 'products.csv' data set from Orginal Data folder
df_prods = pd.read_csv(os.path.join(path, '02 Data','Original Data', 'products.csv'), index_col = False)

# 04. Data Wrangling

## Dropping eval_set column from orders.csv

# Drop the unnecessary column from order data set.
df_ords.drop(columns = ['eval_set'])

# Print all values stored in 'days_since_prior_order' column.
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Import 'department.csv' data set. 
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

#Transposing Data
df_dep.T

#Creating a new transposed df_dep dataframe
df_dep_t = df_dep.T

df_dep_t.head()

df_dep_t

# Creating an index for df_dep_t dataframe
df_dep_t.reset_index()

# Creating a header with the 1st line of df_dep_t dataframe
new_header = df_dep_t.iloc[0]

new_header

# Take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

#Set the header row as the df header
df_dep_t_new.columns = new_header

df_dep_t_new

## Turn the 'department.csv' data set into dictionary.
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

# Checking the data dictionary for what department 19 represent
print(data_dict.get('19'))

# Checking the data dictionary for what department 13 represent
print(data_dict.get('13'))

# 05. Subsetting

df_prods['department_id'] ==19

df_prods[df_prods['department_id'] ==19]

# Create a df_snacks dataframe with only product with 'department_id' of 19
df_snacks = df_prods[df_prods['department_id'] ==19]

df_snacks

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2

# Create a df_snacks_3 dataframe with only product with 'department_id' of 17,18,19
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([17,18,19])]

df_snacks_3

# TASK

## 02. Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.

# Define new dataframe and store the output of dropped column
df_ords_2 = df_ords.drop(columns = ['eval_set'])

df_ords_2

# Changing the data type of column 'order_number' to string
df_ords_2['order_number'] = df_ords_2['order_number'].astype('str')

# Check the data type of dropped column.
df_ords_2['order_number'].dtype

## 03. Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the dataframe.

df_ords_2.head()

# Renaming the variables in data frame. 
df_ords_2.rename(columns = {'order_dow':'order_day_of_week', 'days_since_prior_order' : 'days_since_last_order'}, inplace = True)

df_ords_2.head()

## 04. Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

# checking the busiest hour of the day.
df_ords_2['order_hour_of_day'].value_counts()

## Ans: The busiest hour of the day is 10am with the frequency of 288418 orders. 

## 05. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.


# Finding the value of 4 in 'department_id' column within df_prods dataframe. 
print(data_dict.get('4'))

## Ans: The value of 4 represents 'produce' under department_id column. 

## 06. The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.


# Check the department number that represent 'breakfast'
data_dict


# Create a df_brkfst subset dataframe with only product with 'department_id' of 14
df_brkfst = df_prods[df_prods['department_id'] == 14]

df_brkfst.head()

## 07. They’d also like to see details about products that customers might use to throw dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

# Create a df_dinner_parties dataframe with product with 'department_id' of 5(alcohol), 7(beverages), 12(meat seafood), 20(deli)
df_dinner_parties = df_prods[df_prods['department_id'].isin([5,7,12,20])]

df_dinner_parties.head(20)

# Look for number of each department. 
df_dinner_parties['department_id'].value_counts()

## 08. It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?

# Total counts of rows in dataframes.
df_dinner_parties.shape

## 09. Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.

# Information of customer with 'user_id' of '1'
df_user_1 = df_ords_2.loc[df_ords_2['user_id'] == 1]

df_user_1

df_user_1['user_id'].value_counts()

## 10. You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?


# More details about customer with user_id = 1
df_user_1.describe()

## 12. Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data” folder.

#Exporting df_ords_2 dataframe to prepared Data folder
df_ords_2.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

## 13. Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data” folder so that you have a “.csv” file of your departments data in the correct format.

#Exporting df_dep_t_new dataframe to Prepared Data folder
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
