==> Instacart Grocery Analysis/03 Scripts/4. Exercise and Task Data Consistency Checks 4.5.ipynb <==
# EXERCISE 4.5 

# 01. Import Libraries

# Import pandas and NumPy libraries
import pandas as pd
import numpy as np
import os

# 02. Import Data set from Original Data folder

# Turn the folder path into variable to get access easily. 
path = r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4'

# Import 'products.csv' data set from Orginal Data folder
df_prods = pd.read_csv(os.path.join(path, '02 Data','Original Data', 'products.csv'), index_col = False)

df_prods.head()

# Import order_wrangled.csv data set from Prepared Data folder
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_ords.head()

# 03. Mixed-Type Data

# Create a new Dataframe
df_test = pd.DataFrame()

# Create a mox type column 
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# check for mixed type
for col in df_test.columns.tolist():
  weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# 04. Missing Data

# Find the missing values in products.csv data set
df_prods.isnull().sum()

# To actually view these 16 values, you can create a subset of the dataframe containing only the missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Check the number of rows in your current dataframe
df_prods.shape

# Create a new dataframe of clean data
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean

df_prods_clean.shape

# 05. Duplicate Data

# To check duplicate values
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# create a new dataframe that doesn’t include the duplicates you just identified using the drop_duplicates() function
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# export the dataframe into csv file.
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

df_prods_clean_no_dups.shape

# TASK 4.5

## Step 2.

df_ords.describe()

#### Days_Since_last_Order: The "days_since_last_order" column has a lower count compared to other rows, which might indicate that there are missing values in the data. A value of zero in this column indicates that some orders were placed on the same day. The maximum value is 30, which suggests that there might be a limit on the time window within which orders are recorded.
#### Order_Number: The "order_number" column has a maximum value of 100, which could imply that there is a cap on the number of items that can be ordered per transaction.
#### Order_Day_of_the_Week: The "order_day_of_week" column contains values ranging from 0 to 6, where these numbers correspond to the days of the week. This gives a clear indication of the day on which the order was placed.
#### Order_Hour_of_the_Day: The "order_hour_of_day" column has values from 0 to 23, representing the 24 hours in a day. This gives insight into the specific time of day the order was placed (e.g., 0 = midnight, 12 = noon, 23 = 11 PM).

## Step 3

# Check for mixed-type data in your df_ords dataframe.
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

# There's no mixed-type data in df_ords dataframe.
df_ords.dtypes

## Step 5

# Check for missing data in dataframe.
df_ords.isnull().sum()

### The 'days_since_last_order' column has 206209 missing values. This could happen due to customers' first orders, as these orders wouldn't have any prior order to reference or the missing value for the time gap.

## Step 6

### Address the missing values using an appropriate method.


# To determine the missing values in df_ords dataframe, we need to create a subset dataframe that contains only null values.
df_ords_nan = df_ords[df_ords['days_since_last_order'].isnull() == True]

df_ords_nan

# To check the 'order_number'column with values of 1 are missing values in the 'days_since_last_order' column, 
# then we need to apply filter the dataframe where 'order number' is '1'.
df_orders_1 = df_ords_nan[df_ords_nan['order_number'] == '1']

# Check if all missing values are in 'days_since_last_order' column.
df_orders_1_missing = df_orders_1['days_since_last_order'].isna().all()

# Print the result of missing values in 'days_since_last_order'.
print(df_orders_1_missing) 

# Address the missing values by creating a column that flags order numbers as either True/False depending on if the 'order_number' = 1
df_ords['first_order'] = df_ords['days_since_last_order'].isna()

df_ords

# Check data-type of new column and update if necessary
df_ords['first_order'].dtype

### Now, we have addressed the missing values in days_since_last_order' column appropriately, the next step I would prefer to contact the client to know why these missing values are. However, the data in 'day_since_last_order' column would be helpful for further analysis, so removing or filtering data is not a better choice. 

## Step 7

### Run a check for duplicate values in your df_ords data.

# Check for duplicate values
df_ords_dup = df_ords[df_ords.duplicated()]

df_ords_dup

### There are no duplicate values in df_ords dataframe. 

## Step 9

### Export the cleaned 'df_ords' data set. 

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))
