==> Instacart Grocery Analysis/03 Scripts/5. Exercise Combining and Export Data 4.6.ipynb <==
# EXERCISE 4.6

# Import Libraries

import pandas as pd
import numpy as np
import os

# Import Data Set from Folder

# Turn the folder path into variable to get access easily. 
path = r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4'

# Import order_product_prior.csv from original Data folder in df_ords_prior dataframe
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)


# Import orders_checked.csv from Prepared Data folder.
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_ords_prior.head()

df_ords.head()

df_ords_prior.shape

df_ords.shape

# Merge order data set with order prior data set using order_id column
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# Export pickle file to Prepared Data folder
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
