==> Instacart Grocery Analysis/03 Scripts/6. Exercise 4.7 Deriving New Variable.ipynb <==
# Import Libraries

# Importing pandas and numpy libraries
import pandas as pd
import numpy as np
import os

# Turn folder path into string named as path
path =  r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4'

# Import pickle file from Prepared Data
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

# Create a subset of first million rows
df = df_ords_prods_merge[:1000000]

# Define a function for price label
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply price range function.
df['price_range'] = df.apply(price_label, axis=1)

df.shape

# If-Statements with the loc() Function
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

# Find the frequency of 'order_day_of_week'.
df_ords_prods_merge['order_day_of_week'].value_counts(dropna = False)


result = []

for value in df_ords_prods_merge['order_day_of_week']:
  if value == 0:
    result.append('Busiest day')
  elif value == 4:
    result.append('Least busy')
  else:
    result.append('Regularly busy')

result

df_ords_prods_merge['Busiest Day'] = result

df_ords_prods_merge['Busiest Day'].value_counts(dropna = False)

# Update 'Busiest day' column to 'Busiest days' to identify order day of the week as 'Busiest days', 'Slowest days',  or 'Regularly busy'.

# Create column “busiest_days" from "orders_day_of_week" to identify two busiest days (0, 1), slowest days (4, 3), or regularly busy (all other days).

result_2 = []

for value in df_ords_prods_merge['order_day_of_week']:
  if value == 0 or value == 1:
    result_2.append('Busiest days')
  elif value == 4 or value == 3:
    result_2.append('Slowest days')
  else:
    result_2.append('Regularly busy')

result_2

df_ords_prods_merge.head(50)

# Check the values of 'Busiest Days' for accuracy. Note any observations in markdown format.

# Create new 'busiest_days' column in ords_prods_merge to view results in context.
df_ords_prods_merge['Busiest_days'] = result_2

df_ords_prods_merge['Busiest_days'].value_counts(dropna = False)

The total count of whole dataframe is same as the total sum of grouping columns i.e. Regularly busy, Busiest days and Slowest days. 

# Create a new column named as 'busiest_period_of_day' to identify periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.”

# Check value counts in "order_hour_of_day" column.
df_ords_prods_merge['order_hour_of_day'].value_counts()

# Create for-loop if statement labeling periods of time as “Most orders,” “Average orders,” and “Fewest orders.”
result_3 = []

for value in df_ords_prods_merge["order_hour_of_day"]:
  if value in [10, 11, 14, 15, 13, 12, 16, 9]:
    result_3.append("Most orders")
  elif value in [23, 6, 0, 1, 5, 2, 4, 3]:
    result_3.append("Fewest orders")
  else:
    result_3.append("Average orders")

result_3

# Create new column "busiest_period_of_day" in ords_prods_merge.
df_ords_prods_merge['busiest_period_of_day'] = result_3

# Print the frequency for this new column.

# Print the counts in "busiest_period_of_day" column.
df_ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

# Export your dataframe as a pickle file 

df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))
