==> Instacart Grocery Analysis/03 Scripts/6.1 Task Combining and Export Data 4.6.ipynb <==
# TASK 4.6

# Import Libraries

# Import pandas and numpy libraries
import pandas as pd
import numpy as np
import os

# Create path for file orders_products_combined

path = path = r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4'

# Import orders_products_combined.pkl and create dataframe
orders_product_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))


orders_product_combined.shape

# Import product_cleaned data set
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

# Information about Dataframe

orders_product_combined.head()

# Dropping column 'Unnamed: 0' from dataframe
orders_product_combined = orders_product_combined.drop('Unnamed: 0', axis = 1)

df_prods.shape

df_prods.head()

# Dropping column 'Unnamed: 0' from dataframe
df_prods = df_prods.drop('Unnamed: 0', axis = 1)

# Merging dataframes

### We have a common key in both data sets is 'product_id' so we can say when merging these two Dataframes using the 'product_id' column, an inner join is the most appropriate choice. 

## Step 1:

# First of all, test the merge without overwriting with merge flag 'match'
pd.merge(orders_product_combined , df_prods , on = ['product_id'] , how = 'inner', indicator = 'match')# 

## Step 2: 

# Merge 'orders_product_combined' and 'df_prods'
df_ords_prods_merge = orders_product_combined.merge(df_prods , on = ['product_id'] , indicator = 'match' )

# Now, drop the flag column '_merge' from above merge
df_ords_prods_merge.drop('_merge', axis = 1 )

df_ords_prods_merge.head()

df_ords_prods_merge.shape

## Confirm the results of the merge using the merge flag.

df_ords_prods_merge['_merge'].value_counts()

## Export the newly created dataframe as ords_prods_merge in a suitable format 

df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))
