==> Instacart Grocery Analysis/03 Scripts/7. Task 4.8 Grouping Data & Aggregating Variables.ipynb <==
# Import Libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

## Create a path for folder

# Turn the data folder path into string
path = r'/Users/amritkaur/Desktop/CF Data Immersion/Chapter 4/Achievement 4'

## Import Data from Prepared data folder

# Import data from Prepared Data folder
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))

df_ords_prods_merge.shape

# Grouping and Aggregating Data on entire dataframe

# grouping 'product_name' column using groupby() function
df_ords_prods_merge.groupby('product_name')

# To calculate the mean of the “order_number” column grouped by the “department_id” column.
df_ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

### How do the results for the entire dataframe differ from those of the subset? 
The subset of dataframe contains only certain rows or department IDs, and it represents a specific portion of the data. In this case, for department ID 4, the mean value is higher in the subset (18.825780). In entire dataframe, when we expand the analysis to include the full dataset, the mean value for department ID 4 decreases (17.811403). This suggests that when we incorporate the full data, the values from other department IDs or rows not included in the subset influence the overall mean.
If the full dataset includes some extreme values or outliers that were absent from the subset, these outliers could be distorting the mean in the entire dataframe. For example, if there are some very low values in department 4 that were not in the subset, the overall mean will decrease. If the subset you selected happened to be more representative of higher values for department 4, the mean in the subset could have been biased upward. 

# Create a Loyalty flag

# Aggregating Data with transform() Function
df_ords_prods_merge['max_order'] = df_ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Checking the Output
df_ords_prods_merge.head(15)

# Creating a 'Loyalty' Label Flag using loc Function
df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

df_ords_prods_merge.loc[(df_ords_prods_merge['max_order'] <= 40) & (df_ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Examining Value Counts on 'loyalty_flag'
df_ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

# Checking the Updated Output
df_ords_prods_merge.head(15)

# Analyze the spending habits of Loyalty customers

# Determining Basic Statistics of Spending Habits by Loyalty Category
df_ords_prods_merge.groupby('loyalty_flag').agg({'prices' : ['mean', 'min', 'max']})

As we can see that Loyal customers generally purchases low-cost items with the average of $10.39 and while new customers tend to spend more, averaging $13.29 per item. Regular customers fall between the two, with an average cost of $12.50 per item. 
While all customer groups have a minimum price of $1.00, the presence of an unusually high maximum price of $99,999 indicates that there may be outliers or data errors skewing the results.

# Create a Spending flag

# Aggregating Data with transform() Function
df_ords_prods_merge['avg_product_price'] = df_ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# Checking Relevant Columns 
df_ords_prods_merge[['user_id', 'avg_product_price']].head(100)

# Creating a 'Spending' Label Flag using loc Function
df_ords_prods_merge.loc[df_ords_prods_merge['avg_product_price'] >= 10, 'spending_flag'] = 'High spender'
df_ords_prods_merge.loc[df_ords_prods_merge['avg_product_price'] < 10, 'spending_flag'] = 'Low spender'

# Examining Value Count on 'spending_flag' Column
df_ords_prods_merge['spending_flag'].value_counts(dropna = False)

# Checking Releavent Columns
df_ords_prods_merge[['user_id', 'avg_product_price', 'spending_flag']].head(100)

# Create an order frequency flag

# Aggregating Data with transform() Function
df_ords_prods_merge['median_days_between_orders'] = df_ords_prods_merge.groupby(['user_id'])['days_since_last_order'].transform(np.median)

# Checking Relevant Columns 
df_ords_prods_merge[['user_id', 'days_since_last_order', 'median_days_between_orders']].head(100)

# Creating a 'Order Frequency' Label Flag using loc Function
df_ords_prods_merge.loc[df_ords_prods_merge['median_days_between_orders'] > 20, 'order_frequency_flag'] = 'Non-frequent customer'

df_ords_prods_merge.loc[(df_ords_prods_merge['median_days_between_orders'] <= 20) & (df_ords_prods_merge['median_days_between_orders'] > 10), 'order_frequency_flag'] = 'Regular customer'

df_ords_prods_merge.loc[df_ords_prods_merge['median_days_between_orders'] <= 10, 'order_frequency_flag'] = 'Frequent customer'

# Examining Value Counts on 'order_frequency_flag'
df_ords_prods_merge['order_frequency_flag'].value_counts(dropna = False)

# Checking Relevant Columns of Updated Dataframe
df_ords_prods_merge[['user_id', 'days_since_last_order', 'median_days_between_orders', 'order_frequency_flag']].head(100)

# Export the Dataframe in pickle Format to the "Prepared Data" Folder

# Exporting the  updated dataframe
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_grouped.pkl'))
